Pipeline: reorganize template engine code
